fix: ghsa-8gj5-r8vm-mghq

+++ interface/main/messages/messages.php
@@ -332,6 +332,10 @@ if (!empty($_REQUEST['go'])) { ?>
                             $noteid = $_POST['noteid'];
                             $form_message_status = $_POST['form_message_status'];
                             $reply_to = $_POST['reply_to'];
+                            // IDOR protection: verify note is assigned to current user
+                            if (!checkPnotesNoteId($noteid, $_SESSION['authUser'])) {
+                                die("Message is not assigned to you. Update is disallowed.");
+                            }
                             if ($task == "save") {
                                 updatePnoteMessageStatus($noteid, $form_message_status);
                             } else {
@@ -370,6 +374,10 @@ if (!empty($_REQUEST['go'])) { ?>
                             // Delete selected message(s) from the Messages box (only).
                             $delete_id = $_POST['delete_id'];
                             for ($i = 0; $i < count($delete_id); $i++) {
+                                // IDOR protection: verify note is assigned to current user
+                                if (!checkPnotesNoteId($delete_id[$i], $_SESSION['authUser'])) {
+                                    continue; // skip notes not assigned to user
+                                }
                                 deletePnote($delete_id[$i]);
                                 EventAuditLogger::getInstance()->newEvent("delete", $_SESSION['authUser'], $_SESSION['authProvider'], 1, "pnotes: id " . $delete_id[$i]);
                             }
